adding batching to time, direction and earliest queries
