+++ src/main/java/com/flightstats/hub/channel/ChannelLatestResource.java
@@ -2,23 +2,15 @@ package com.flightstats.hub.channel;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.flightstats.hub.dao.ChannelService;
-import com.flightstats.hub.dao.Request;
-import com.flightstats.hub.model.Content;
 import com.flightstats.hub.model.ContentKey;
 import com.flightstats.hub.model.DirectionQuery;
 import com.google.common.base.Optional;
-import com.google.common.io.ByteStreams;
 import com.google.inject.Inject;
-import org.apache.commons.lang3.RandomStringUtils;
 
 import javax.ws.rs.*;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
-import javax.ws.rs.core.StreamingOutput;
 import javax.ws.rs.core.UriInfo;
-import java.io.BufferedOutputStream;
-import java.io.IOException;
-import java.io.OutputStream;
 import java.net.URI;
 import java.util.Collection;
 
@@ -28,7 +20,6 @@ import static javax.ws.rs.core.Response.Status.SEE_OTHER;
 @Path("/channel/{channel: .*}/latest")
 public class ChannelLatestResource {
 
-    public static final String CRLF = "\r\n";
     @Inject
     private UriInfo uriInfo;
     @Inject
@@ -83,43 +74,8 @@ public class ChannelLatestResource {
         query.trace(trace);
         Collection<ContentKey> keys = channelService.getKeys(query);
         keys.add(latest.get());
-        String boundary = RandomStringUtils.randomAlphanumeric(70);
         if (batch) {
-            //todo - gfm - 8/19/15 - factor this out
-            StreamingOutput stream = new StreamingOutput() {
-                @Override
-                public void write(OutputStream os) throws IOException,
-                        WebApplicationException {
-                    URI channelUri = LinkBuilder.buildChannelUri(channel, uriInfo);
-                    BufferedOutputStream output = new BufferedOutputStream(os);
-                    for (ContentKey key : keys) {
-                        Request request = Request.builder()
-                                .channel(channel)
-                                .key(key)
-                                .build();
-                        Optional<Content> content = channelService.getValue(request);
-                        if (content.isPresent()) {
-                            Content item = content.get();
-                            output.write(("--" + boundary + CRLF).getBytes());
-                            if (item.getContentType().isPresent()) {
-                                output.write(("Content-Type: " + item.getContentType().get() + CRLF).getBytes());
-                            }
-                            URI uri = LinkBuilder.buildItemUri(key, channelUri);
-                            output.write(("Content-Key: " + uri.toString() + CRLF).getBytes());
-                            output.write(CRLF.getBytes());
-                            ByteStreams.copy(item.getStream(), output);
-                            output.write(CRLF.getBytes());
-                            output.flush();
-                        }
-                    }
-                    output.write(("--" + boundary + "--").getBytes());
-                    output.flush();
-                }
-            };
-            Response.ResponseBuilder builder = Response.ok(stream);
-            builder.type("multipart/mixed; boundary=" + boundary);
-
-            return builder.build();
+            return MultiPartBuilder.build(keys, channel, channelService, uriInfo);
         } else {
             return LinkBuilder.directionalResponse(channel, keys, count, query, mapper, uriInfo, true);
         }
